refactor(py/node): split execute_node into program, command_dispatch, classify_argv
